% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data.R
\name{check_data}
\alias{check_data}
\title{Check the data}
\usage{
check_data(data_source_check, silent = FALSE)
}
\arguments{
\item{data_source_check}{List with \code{community} and \code{age}}

\item{silent}{Logical. If \code{TRUE}, suppress all console outputs. Useful for testing.}
}
\description{
Output summary information about the data
}
\keyword{internal}
