% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_data.R
\name{extract_data}
\alias{extract_data}
\title{Extract and check the data}
\usage{
extract_data(
  data_community_extract,
  data_age_extract,
  age_uncertainty = NULL,
  verbose = FALSE,
  silent = FALSE
)
}
\arguments{
\item{data_community_extract}{Data.frame. Community data with species as columns and
levels (samples) as rows. First column should be \code{sample_id} (character).}

\item{data_age_extract}{Data.frame with two columns:
\itemize{
\item \code{sample_id} - unique ID of each level (character)
\item \code{age} - age of level (numeric)
}}

\item{age_uncertainty}{Usage of age uncertainty form Age-depth models. Either:
\itemize{
\item matrix with number of columns as number of samples. Each column is one sample,
each row is one age sequence from age-depth model. Age sequence is randomly
sampled from age-depth model uncertainties at the beginning of each run.
\item \code{NULL} - Age uncertainties are not available and, therefore, will not be used.
}}

\item{verbose}{\code{logical}. If \code{TRUE}, print progress messages during extraction.}

\item{silent}{Logical. If \code{TRUE}, suppress all console outputs (overrides verbose). Useful for testing.}
}
\value{
A named \code{list} with elements \code{community} (\code{data.frame}), \code{age}
(\code{data.frame}), and optionally \code{age_un} (\code{matrix}).
}
\description{
Extract, validate, and align the community and age data ready for
RoC estimation.
}
\keyword{internal}
