% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_check_data.R
\name{fc_check_data}
\alias{fc_check_data}
\title{Check the data}
\usage{
fc_check_data(data_source_check)
}
\arguments{
\item{data_source_check}{List with \code{community} and \code{age}}
}
\description{
Output summary information about the data
}
