% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_make_bins.R
\name{fc_make_bins}
\alias{fc_make_bins}
\title{Create template for RoC estimations}
\usage{
fc_make_bins(
  data_source_bins,
  Working_Units = c("levels", "bins", "MW"),
  bin_size = 500,
  Number_of_shifts = 5
)
}
\arguments{
\item{data_source_bins}{List with \code{community} and \code{age}}

\item{Working_Units}{Character. Selection of units that the DC will be calculated between.
\itemize{
\item \code{"levels"} - individual levels are going to be used
\item \code{"bins"} - samples in predefined bins will be pooled together and one sample
will be selected from each time bin as a representation.
\item \code{"MW"} - Bins of selected size are created, starting from the beginning of the core.
This is repeated many times, with each time bin (window) shifting by Z years forward.
This is repeated X times, where X = bin size / Z.
}}

\item{bin_size}{Numeric. Size of the time bin (in years)}

\item{Number_of_shifts}{Numeric. Value determining the number of shifts of window used
in Moving window method}
}
\description{
Create template for RoC estimations
}
