% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_reduce.R
\name{fc_reduce}
\alias{fc_reduce}
\title{Reduce datasets}
\usage{
fc_reduce(data_source_reduce, check_taxa = TRUE, check_levels = TRUE)
}
\arguments{
\item{data_source_reduce}{List with \code{community}, \code{age}, and \code{age_un}}

\item{check_taxa}{Logical. Should columns be check for redundnat data?}

\item{check_levels}{Logical. Should rows be check for redundnat data?}
}
\description{
Check the community dataset for redundnat taxa and levels
and filter them out.
}
\keyword{internal}
