% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_smooth_community_data.R
\name{fc_smooth_community_data}
\alias{fc_smooth_community_data}
\title{Smooth the community data}
\usage{
fc_smooth_community_data(
  data_source_smooth,
  smooth_method = c("m.avg", "grim", "age.w", "shep"),
  smooth_N_points = 5,
  smooth_N_max = 9,
  smooth_age_range = 500,
  round_results = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{data_source_smooth}{List with \code{community}, and \code{age}}

\item{smooth_method}{Character. type of smoothing applied for the each of the pollen type
\itemize{
\item \code{"m.avg"} - Moving average
\item \code{"grim"} - Grimm's smoothing
\item \verb{"age.w""} - Age-weighted average
\item \code{"shep"} - Shepard's 5-term filter
}}

\item{smooth_N_points}{Numeric. Number of points for used for moving average,
Grimm and Age-Weighted smoothing (odd number)}

\item{smooth_N_max}{Numeric. Maximal number of samples to look in Grimm smoothing}

\item{smooth_age_range}{Numeric. Maximal age range for both Grimm and Age-weight smoothing}

\item{round_results}{Logical. Should smoothed values be rounded to integers?}

\item{verbose}{Logical. If \code{TRUE}, function will output messages about internal processes}
}
\description{
A function to apply one of the 4 smoothers.
}
