% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_subset_samples.R
\name{fc_subset_samples}
\alias{fc_subset_samples}
\title{Subsetting levels in each Working units (WU)}
\usage{
fc_subset_samples(
  data_source_subset,
  data_source_bins,
  bin_selection = "first"
)
}
\arguments{
\item{data_source_subset}{Data.frame with community data}

\item{data_source_bins}{Data.frame with individual WU to use}

\item{bin_selection}{Character. Setting determining the the process of selection of
samples from bins.
\itemize{
\item \code{"first"} - sample closest to the beginning of the bin is selected
as a representation.
\item \code{"random"} - a random sample is selected as a representation.
}}
}
\description{
Subsetting levels in each Working units (WU)
}
\keyword{internal}
