% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_transfer_into_proportions.r
\name{fc_transfer_into_proportions}
\alias{fc_transfer_into_proportions}
\title{Transform community data into proportions}
\usage{
fc_transfer_into_proportions(
  data_source,
  sel_method = c("proportions", "percentages"),
  verbose = FALSE
)
}
\arguments{
\item{data_source}{List with \code{community} data.frame with pollen data.
Each row represent one level (sample) and each column represent one taxon.
able must contain sample_id` as rownames.}

\item{sel_method}{variable to select result as either proportions (\code{percentages}) or
percentage (\code{percentages}).}
}
\description{
Tranform pollen data into proportions (or percentages)
}
