% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl_bchron.R
\docType{data}
\name{gl_bchron}
\alias{gl_bchron}
\title{Bchron result used in the vignette Example of full workflow"}
\format{
A list of 12:
\describe{
 \item{theta}{}
 \item{phi}{} 
 \item{mu}{}
 \item{thetaPredict}{}
 \item{predictPositions}{}
 \item{positions}{}
 \item{extractDate}{}
 \item{ageScaleVal}{}
 \item{positionNormalise}{}
 \item{inputVals}{}
}
}
\usage{
gl_bchron
}
\description{
Bchron result used in the vignette Example of full workflow"
}
\examples{
\dontrun{
 gl_bchron
}
}
\keyword{datasets}
