% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl_roc.R
\docType{data}
\name{gl_roc}
\alias{gl_roc}
\title{Result of RoC estimates used in the vignette Example of full workflow"}
\format{
A tibble with 5 columns:
\describe{
 \item{Working_Unit}{}
 \item{Age   }{} 
 \item{ROC }{}
 \item{ROC_up}{}
 \item{ROC_dw}{}
}
}
\usage{
gl_roc
}
\description{
Result of RoC estimates used in the vignette Example of full workflow"
}
\examples{
\dontrun{
 gl_roc
}
}
\keyword{datasets}
