% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce_data.R
\name{reduce_data}
\alias{reduce_data}
\title{Reduce datasets}
\usage{
reduce_data(data_source_reduce, check_taxa = TRUE, check_levels = TRUE)
}
\arguments{
\item{data_source_reduce}{\code{list} with \code{community}, \code{age}, and \code{age_un}}

\item{check_taxa}{\code{logical}. Should columns be checked for all-zero taxa?}

\item{check_levels}{\code{logical}. Should rows be checked for empty levels?}
}
\value{
The input \code{list} with redundant columns and rows removed.
}
\description{
Check the community dataset for all-zero taxa and empty levels and
filter them out.
}
\keyword{internal}
