% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce_data_simple.R
\name{reduce_data_simple}
\alias{reduce_data_simple}
\title{Reduce datasets in merged format}
\usage{
reduce_data_simple(
  data_source_reduce,
  ommit_vars = c("label", "res_age", "age_diff"),
  check_taxa = TRUE,
  check_levels = TRUE
)
}
\arguments{
\item{data_source_reduce}{\code{list} with \code{community}, \code{age}, and \code{age_un}}

\item{ommit_vars}{\code{character} vector of column names to exclude from community data.}

\item{check_taxa}{\code{logical}. Should columns be checked for all-zero taxa?}

\item{check_levels}{\code{logical}. Should rows be checked for empty levels?}
}
\value{
The input \code{data.frame} with redundant columns and rows removed.
}
\description{
Check the community dataset for all-zero taxa and empty levels and
filter them out. This simplified version is optimised for performance.
}
\keyword{internal}
