% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_community_data.R
\name{smooth_community_data}
\alias{smooth_community_data}
\title{Smooth the community data}
\usage{
smooth_community_data(
  data_source_smooth,
  smooth_method = c("m.avg", "grim", "age.w", "shep"),
  smooth_n_points = 5,
  smooth_n_max = 9,
  smooth_age_range = 500,
  round_results = FALSE,
  verbose = FALSE,
  silent = FALSE
)
}
\arguments{
\item{data_source_smooth}{List with \code{community}, and \code{age}}

\item{smooth_method}{Character. type of smoothing applied for the each of the pollen type
\itemize{
\item \code{"none"} - Pollen data is not smoothed
\item \code{"m.avg"} - Moving average
\item \code{"grim"} - Grimm's smoothing
\item \verb{"age.w""} - Age-weighted average
\item \code{"shep"} - Shepard's 5-term filter
}}

\item{smooth_n_points}{Numeric. Number of points for used for moving average,
Grimm and Age-Weighted smoothing (odd number)}

\item{smooth_n_max}{Numeric. Maximal number of samples to look in Grimm smoothing}

\item{smooth_age_range}{Numeric. Maximal age range for both Grimm and Age-weight smoothing}

\item{round_results}{Logical. Should smoothed values be rounded to integers?}

\item{verbose}{Logical. If \code{TRUE}, function will output messages about internal processes}

\item{silent}{Logical. If \code{TRUE}, suppress all console outputs (overrides verbose). Useful for testing.}
}
\description{
A function to apply one of the 4 smoothers.
}
\details{
Smoothing of assemblage data: Each variable within the
assemblage data is smoothed using one of five in-built smoothing methods:
\itemize{
\item Shepard's 5-term filter (\code{smooth_method} = \code{"shep"};
Davis, 1986; Wilkinson, 2005)
\item moving average (\code{smooth_method} = \verb{"m.avg"\}})
\item age-weighted average (\code{smooth_method} = \code{"age.w"})
\item Grimm's smoothing (\code{smooth_method} = \code{"grim"};
Grimm and Jacobson, 1992)
}
}
\references{
Davis, J.C., 1986. Statistics and Data Analysis in Geology, 2nd edn. ed.
J. Wiley & Sons, New York.

Grimm, E.C., Jacobson, G.L., 1992. Fossil-pollen evidence for abrupt
climate changes during the past 18000 years in eastern North America.
Clim. Dyn. 6, 179-184.

Wilkinson, L., 2005. The Grammar of Graphics. Springer-Verlag, New York,
USA 37.
}
\seealso{
\code{\link[=estimate_roc]{estimate_roc()}}
}
