% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_community.R
\name{subset_community}
\alias{subset_community}
\title{Subset data to only contain community (taxa)}
\usage{
subset_community(
  data_source,
  ommit_vars = c("label", "res_age", "age_diff", "age")
)
}
\arguments{
\item{data_source}{\code{data.frame} to clean.}

\item{ommit_vars}{\code{character} vector of column names to exclude.}
}
\value{
A \code{data.frame} containing only taxon columns.
}
\description{
Subset data to only contain community (taxa)
}
\keyword{internal}
