% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_samples.R
\name{subset_samples}
\alias{subset_samples}
\title{Subset levels into Working Units}
\usage{
subset_samples(data_source_subset, data_source_bins, bin_selection = "first")
}
\arguments{
\item{data_source_subset}{\code{data.frame} with community data.}

\item{data_source_bins}{\code{data.frame} defining individual Working Units.}

\item{bin_selection}{\code{character}. Rule for selecting one level from each bin.
\itemize{
\item \code{"random"} (default) - a level is selected at random.
\item \code{"first"} - the level closest to the start of the bin is
selected.
}}
}
\value{
A \code{data.frame} with one row per Working Unit.
}
\description{
Select one representative level per bin according to \code{bin_selection}.
}
\keyword{internal}
