% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_check_data_assembly.R
\name{util_check_data_assembly}
\alias{util_check_data_assembly}
\title{Check the data assembly}
\usage{
util_check_data_assembly(data_source)
}
\arguments{
\item{data_source}{Data.frame to be tested}
}
\value{
Data.frame with all variables
}
\description{
Check the dataset for mandatory and non-mandatory
variables. Stop the session if mandatory not present. Warn about
non-mandatory and fill them with 'NA'
}
