% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_check_if_integer.R
\name{util_check_if_integer}
\alias{util_check_if_integer}
\title{Assert the class of object}
\usage{
util_check_if_integer(data_source)
}
\arguments{
\item{data_source}{The name of the object in quotes}
}
\value{
\code{TRUE} if class match or error message
}
\description{
The function will evaluate the object of \code{data_source} name
and test if ANY of the classes in equal to \code{sel_class}
}
