% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_check_vector_values.R
\name{util_check_vector_values}
\alias{util_check_vector_values}
\title{Check the presence of variables in vector}
\usage{
util_check_vector_values(data_source, var_list)
}
\arguments{
\item{data_source}{Name of the vector to check}

\item{var_list}{Vector with names}
}
\description{
Function will test the presence of ANY values from \code{var_list}
within the \code{data_source} vector and return error message
}
