% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_flatten_by_one.R
\name{util_flatten_by_one}
\alias{util_flatten_by_one}
\title{Flatten the list by one levles}
\usage{
util_flatten_by_one(x)
}
\arguments{
\item{x}{Nested list}
}
\description{
Return list without the first h-level. The names of the h1 and
h2 are merged
}
