% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_output_comment.R
\name{util_output_comment}
\alias{util_output_comment}
\title{Output message in console}
\usage{
util_output_comment(msg = "")
}
\arguments{
\item{msg}{String with the message that should be printed}
}
\description{
Output message in console
}
