% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_output_message.R
\name{util_output_message}
\alias{util_output_message}
\title{Output boxed message in console}
\usage{
util_output_message(msg = "")
}
\arguments{
\item{msg}{String with the message that should be printed}
}
\description{
Output boxed message in console
}
