% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_output_warning.R
\name{util_output_warning}
\alias{util_output_warning}
\title{Output warning in console}
\usage{
util_output_warning(msg = "")
}
\arguments{
\item{msg}{String with the message that should be printed}
}
\description{
Output warning in console
}
