% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data.R
\name{check_data}
\alias{check_data}
\title{Check the data}
\usage{
check_data(data_source_check, silent = FALSE)
}
\arguments{
\item{data_source_check}{\code{list} with \code{community} (\code{data.frame}) and \code{age} (\code{data.frame})
elements.}

\item{silent}{\code{logical}. If \code{TRUE}, suppress all console output.}
}
\value{
\code{NULL} invisibly.
}
\description{
Output summary information about the input data to the console.
}
\keyword{internal}
