% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_sni.R
\name{detect_sni}
\alias{detect_sni}
\title{Calculate signal-to-noise index (SNI) for a charcoal record}
\usage{
detect_sni(CharData, BandWidth)
}
\arguments{
\item{CharData}{Matrix of input data with one row per sample, containing:
\itemize{
\item Column 1: age associated with the sample (yr)
\item Column 2: charcoal accumulation rate (CHAR) of the sample
(pieces/cm^2/yr)
\item Column 3: threshold value (pieces/cm^2/yr)
}}

\item{BandWidth}{= Width of moving window for computing SNI}
}
\value{
SNI_output is a data list containing the computed SNI and related data,
with one row for each row in the input variable CharData:
\itemize{
\item SNI = the SNI computed for each sample
\item winInd = indexes of the first and last samples included in each moving
window. E.g. \code{SNI_output["winInd", X]} == (A, B) indicates that the
moving window used to calculate SNI for the Xth sample contained all
samples between A and B, inclusive.
\item popN = the CHAR values of all samples in the noise (N) population
(samples in the moving window with CHAR below threshold)
\item popS = the CHAR values of all samples in the signal (S) population
(samples in the moving window with CHAR at or above threshold)
\item meanN = mean CHAR of the samples in popN
\item stdN = standard deviation of the samples in popN
\item CF = the "correction factor" used in computing SNI.  Equal to
(v - 2)/v ,where v is the number of samples in popN
}
}
\description{
Code obtained and from Kelly et al. 2010
}
\details{
This function computes SNI as described in Kelly et al. 2010.
Note that your data must be interpolated to constant sample resolution
(yr/sample) before input to the function. The function makes no assumption
about prior analysis on the input CHAR series, i.e. any background and
threshold methods may be used.  However, input data should still be
consistent with the interpretation that a CHAR value (column 2) greater
than the corresponding threshold value (column 3) is a "signal" sample,
whereas a CHAR value below the threshold is "noise".
Refer to Kelly et al. 2010 for details and discussion.
}
\references{
Kelly RF, Higuera PE, Barrett CM, Feng Sheng H. 2011 A signal-to-noise index
to quantify the potential for peak detection in sediment-charcoal records.
Quat. Res. 75, 11-17. (doi:10.1016/j.yqres.2010.07.011)
}
\keyword{internal}
