% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_dissimilarity_coefficient.R
\name{estimate_dissimilarity_coefficient}
\alias{estimate_dissimilarity_coefficient}
\title{Calculate the dissimilarity coeficient}
\usage{
estimate_dissimilarity_coefficient(
  data_source_dc,
  dissimilarity_coefficient = "chord",
  verbose = FALSE
)
}
\arguments{
\item{data_source_dc}{Data.frame with taxons as columns}

\item{dissimilarity_coefficient}{Character. Dissimilarity coefficient. Type of calculation of differences
between Working Units. See \code{vegan::vegdist} for more details.
\itemize{
\item \code{"euc"} - Euclidean distance
\item \code{"euc.sd"} - Standardised Euclidean distance
\item \code{"chord"} - Chord distance
\item \code{"chisq"} - Chi-squared coefficient
\item \code{"gower"} - Gower's distance
\item \code{"bray"} - Bray-Curtis distance
}}

\item{verbose}{Logical. If \code{TRUE}, function will output messages about internal processes}
}
\description{
Calculate the dissimilarity coeficient
}
\details{
Five in-built dissimilarity coefficients are available:
\itemize{
\item Euclidean distance (\code{dissimilarity_coefficient} = \code{"euc"})
\item standardised Euclidean distance (\code{dissimilarity_coefficient} = \code{"euc.sd"})
\item Chord distance (\code{dissimilarity_coefficient} = \code{"chord"})
\item Chi-squared coefficient (\code{dissimilarity_coefficient} = \code{"chisq"})
\item Gower's distance (\code{dissimilarity_coefficient} = \code{"gower"})
\item Bray-Curtis distance (\code{dissimilarity_coefficient} = \code{"bray"})
}
The choice of dissimilarity_coefficient depends on the type of assemblage data. In addition, RoC
between WUs be calculated using every consecutive WU (\code{only_subsequent} = \code{FALSE}),
or alternatively, calculation of RoC can be restricted to only directly
adjacent WUs (\code{only_subsequent} = \code{TRUE}). Using the former increases
the number of samples for which RoC can be calculated within a sequence,
which varies in terms of sample resolution, but may still introduce
biases related to the RoC estimation as a result of the varying
inter-sample distances.
}
\seealso{
\code{\link[vegan:vegdist]{vegan::vegdist()}}
}
