% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_data.R
\name{extract_data}
\alias{extract_data}
\title{Extract and check the data}
\usage{
extract_data(
  data_community_extract,
  data_age_extract,
  age_uncertainty = NULL,
  verbose = FALSE,
  silent = FALSE
)
}
\arguments{
\item{data_community_extract}{Data.frame. Community data with species as columns and
levels (samples) as rows. First column should be \code{sample_id} (character).}

\item{data_age_extract}{Data.frame with two columns:
\itemize{
\item \code{sample_id} - unique ID of each level (character)
\item \code{age} - age of level (numeric)
}}

\item{age_uncertainty}{Usage of age uncertainty form Age-depth models. Either:
\itemize{
\item matrix with number of columns as number of samples. Each column is one sample,
each row is one age sequence from age-depth model. Age sequence is randomly
sampled from age-depth model uncertainties at the beginning of each run.
\item \code{NULL} - Age uncertainties are not available and, therefore, will not be used.
}}

\item{verbose}{DESCRIPTION.
Logical. If \code{TRUE}, function will output messages about internal processes}

\item{silent}{Logical. If \code{TRUE}, suppress all console outputs (overrides verbose). Useful for testing.}
}
\description{
Function for general preparation of input data
}
