% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_detect_peak_points.R
\name{fc_detect_peak_points}
\alias{fc_detect_peak_points}
\title{Detect significant peak points}
\usage{
fc_detect_peak_points(
  data_source,
  sel_method = c("trend_linear", "trend_non_linear", "threshold", "GAM_deriv", "SNI"),
  sd_threshold = 2
)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function was deprecated, please use \code{\link[=detect_peak_points]{detect_peak_points()}}.
}
\seealso{
\code{\link[=detect_peak_points]{detect_peak_points()}}
}
\keyword{internal}
