% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_estimate_RoC.R
\name{fc_estimate_RoC}
\alias{fc_estimate_RoC}
\title{RRatepol: Estimate rate of change}
\usage{
fc_estimate_RoC(
  data_source_community,
  data_source_age,
  age_uncertainty = NULL,
  smooth_method = c("none", "m.avg", "grim", "age.w", "shep"),
  smooth_n_points = 5,
  smooth_N_points = lifecycle::deprecated(),
  smooth_age_range = 500,
  smooth_n_max = 9,
  smooth_N_max = lifecycle::deprecated(),
  working_units = c("levels", "bins", "MW"),
  Working_Units = lifecycle::deprecated(),
  bin_size = 500,
  number_of_shifts = 5,
  Number_of_shifts = lifecycle::deprecated(),
  bin_selection = c("random", "first"),
  standardise = FALSE,
  n_individuals = 150,
  N_individuals = lifecycle::deprecated(),
  dissimilarity_coefficient = c("euc", "euc.sd", "chord", "chisq", "gower", "bray"),
  DC = lifecycle::deprecated(),
  tranform_to_proportions = TRUE,
  rand = NULL,
  use_parallel = FALSE,
  interest_threshold = NULL,
  time_standardisation = NULL,
  verbose = FALSE
)
}
\arguments{
\item{data_source_community}{Data.frame. Community data with species as columns and
levels (samples) as rows. First column should be \code{sample_id} (character).}

\item{data_source_age}{Data.frame with two columns:
\itemize{
\item \code{sample_id} - unique ID of each level (character)
\item \code{age} - age of level (numeric)
}}

\item{age_uncertainty}{Usage of age uncertainty form Age-depth models. Either:
\itemize{
\item matrix with number of columns as number of samples. Each column is one sample,
each row is one age sequence from age-depth model. Age sequence is randomly
sampled from age-depth model uncertainties at the beginning of each run.
\item \code{NULL} - Age uncertainties are not available and, therefore, will not be used.
}}

\item{smooth_method}{Character. type of smoothing applied for the each of the pollen type
\itemize{
\item \code{"none"} - Pollen data is not smoothed
\item \code{"m.avg"} - Moving average
\item \code{"grim"} - Grimm's smoothing
\item \verb{"age.w""} - Age-weighted average
\item \code{"shep"} - Shepard's 5-term filter
}}

\item{smooth_n_points}{Numeric. Number of points for used for moving average,
Grimm and Age-Weighted smoothing (odd number)}

\item{smooth_N_points}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} \code{smooth_N_points} is no
longer supported; please use \code{smooth_n_points}}

\item{smooth_age_range}{Numeric. Maximal age range for both Grimm and Age-weight smoothing}

\item{smooth_n_max}{Numeric. Maximal number of samples to look in Grimm smoothing}

\item{smooth_N_max}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} \code{smooth_N_max} is no
longer supported; please use \code{smooth_n_max}}

\item{working_units}{Character. Selection of units that the dissimilarity_coefficient will be calculated between.
\itemize{
\item \code{"levels"} - individual levels are going to be used
\item \code{"bins"} - samples in predefined bins will be pooled together and one sample
will be selected from each time bin as a representation.
\item \code{"MW"} - Bins of selected size are created, starting from the beginning of the core.
This is repeated many times, with each time bin (window) shifting by Z years forward.
This is repeated X times, where X = bin size / Z.
}}

\item{Working_Units}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} \code{Working_Units} is no
longer supported; please use \code{working_units}}

\item{bin_size}{Numeric. Size of the time bin (in years)}

\item{number_of_shifts}{Numeric. Value determining the number of shifts of window used
in Moving window method}

\item{Number_of_shifts}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} \code{Number_of_shifts} is no
longer supported; please use \code{number_of_shifts}}

\item{bin_selection}{Character. Setting determining the the process of selection of
samples from bins.
\itemize{
\item \code{"first"} - sample closest to the beginning of the bin is selected
as a representation.
\item \code{"random"} - a random sample is selected as a representation.
}}

\item{standardise}{Logical. If \code{standardise} == \code{TRUE}, then standardise
each Working Unit to certain number of individuals (using random resampling
without repetition)}

\item{n_individuals}{Numeric. Number of grain to perform standardisation to.
The \code{N_individual} is automatically adjusted to the smallest number
of pollen grains in sequence.}

\item{N_individuals}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} \code{N_individuals} is no
longer supported; please use \code{n_individuals}}

\item{dissimilarity_coefficient}{Character. Dissimilarity coefficient. Type of calculation of differences
between Working Units. See \code{vegan::vegdist} for more details.
\itemize{
\item \code{"euc"} - Euclidean distance
\item \code{"euc.sd"} - Standardised Euclidean distance
\item \code{"chord"} - Chord distance
\item \code{"chisq"} - Chi-squared coefficient
\item \code{"gower"} - Gower's distance
\item \code{"bray"} - Bray-Curtis distance
}}

\item{DC}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} \code{DC} is no
longer supported; please use \code{dissimilarity_coefficient}}

\item{tranform_to_proportions}{Logical. Should the community data be transformed to a
proportion during calculations?}

\item{rand}{Numeric. Number of runs used in randomisation.}

\item{use_parallel}{Preference of usage of parallel computation of randomisation
\itemize{
\item \verb{[value]} - selected number of cores
\item \code{TRUE} - automatically selected number of cores
\item \code{FALSE} - does not use parallel computation (only single core)
}}

\item{interest_threshold}{Numeric. Optional. Age, after which all results of RoC are excluded.}

\item{time_standardisation}{Numeric. Units scaling for result RoC values. For example,
if \code{time_standardisation} = 100, the RoC will be reported as
dissimilarity per 100 yr.}

\item{verbose}{Logical. If \code{TRUE}, function will output messages about internal processes}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function was deprecated, please use \code{\link[=estimate_roc]{estimate_roc()}}.
}
\seealso{
\code{\link[=estimate_roc]{estimate_roc()}}
}
\keyword{internal}
