% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_trend.R
\name{make_trend}
\alias{make_trend}
\title{Predict data for certain trend}
\usage{
make_trend(data_source, sel_method = c("linear", "non_linear"))
}
\arguments{
\item{data_source}{Data.frame with \code{ROC} and \code{Age}}

\item{sel_method}{Which trend should be used:
\itemize{
\item \code{"linear"} - A linear model is fitted between the RoC values and
their ages.
\item \code{"non_linear"} - A conservative generalised additive model (GAM)
is fitted through the RoC scores and their ages (GAM = \code{RoC ~ s(age, k = 3)}
using the \code{mgcv} package (Wood, 2011).
}}
}
\description{
Predict data for certain trend
}
\seealso{
\code{\link[=detect_peak_points]{detect_peak_points()}}
}
\keyword{internal}
