% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{prepare_data}
\alias{prepare_data}
\title{Prepare data for single RoC estimation}
\usage{
prepare_data(
  data_source_prep,
  working_units = c("levels", "bins", "MW"),
  bin_size = 500,
  number_of_shifts = 5,
  rand = NULL
)
}
\arguments{
\item{data_source_prep}{List with \code{community} and \code{age}}

\item{working_units}{Character. Selection of units that the dissimilarity_coefficient will be calculated between.
\itemize{
\item \code{"levels"} - individual levels are going to be used
\item \code{"bins"} - samples in predefined bins will be pooled together and one sample
will be selected from each time bin as a representation.
\item \code{"MW"} - Bins of selected size are created, starting from the beginning of the core.
This is repeated many times, with each time bin (window) shifting by Z years forward.
This is repeated X times, where X = bin size / Z.
}}

\item{bin_size}{Numeric. Size of the time bin (in years)}

\item{number_of_shifts}{Numeric. Value determining the number of shifts of window used
in Moving window method}

\item{rand}{Numeric. Number of runs used in randomisation.}
}
\description{
Create a single list with all information needed to estimate RoC.
This is done because such list can be then evaluated in parallel.
}
\keyword{internal}
