% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce_data_simple.R
\name{reduce_data_simple}
\alias{reduce_data_simple}
\title{Reduce datasets in merged format}
\usage{
reduce_data_simple(
  data_source_reduce,
  ommit_vars = c("label", "res_age", "age_diff"),
  check_taxa = TRUE,
  check_levels = TRUE
)
}
\arguments{
\item{data_source_reduce}{List with \code{community}, \code{age}, and \code{age_un}}

\item{ommit_vars}{Character vector with names of columns to omit in community data.}

\item{check_taxa}{Logical. Should columns be check for redundnat data?}

\item{check_levels}{Logical. Should rows be check for redundnat data?}
}
\description{
Check the community dataset for redundnat taxa and levels
and filter them out. This function is simplified due to performance.
}
\keyword{internal}
