% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRatepol-package.R
\docType{package}
\name{RRatepol-package}
\alias{RRatepol}
\alias{RRatepol-package}
\title{RRatepol: R-Ratepol: Estimating Rate of Change in Temporal Sequences}
\description{
R-Ratepol is an R package for estimating rate of change (RoC) from community data in time series.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/HOPE-UIB-BIO/R-Ratepol-package}
  \item \url{https://hope-uib-bio.github.io/R-Ratepol-package/}
}

}
\author{
\strong{Maintainer}: Ondrej Mottl \email{ondrej.mottl@gmail.com}

Other contributors:
\itemize{
  \item Friederike Wolke \email{friederike.woelke@gmail.com} [contributor]
}

}
\keyword{internal}
