% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardise_community_data.R
\name{standardise_community_data}
\alias{standardise_community_data}
\title{Standardise the community data}
\usage{
standardise_community_data(data_source_standard, n_individuals = 150)
}
\arguments{
\item{data_source_standard}{Data.frame with taxons as columns}

\item{n_individuals}{Numeric. Number of grain to perform standardisation to.
The \code{N_individual} is automatically adjusted to the smallest number
of pollen grains in sequence.}
}
\description{
Taxa in the assemblage dataset can be standardised to a certain count
(e.g. number of pollen grains in each WU) by rarefaction. Random sampling
without replacement is used to draw a selected number of individuals from
each WU (e.g. 150 pollen grains).
}
