% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_community.R
\name{subset_community}
\alias{subset_community}
\title{Subset data to only contain community (taxa)}
\usage{
subset_community(
  data_source,
  ommit_vars = c("label", "res_age", "age_diff", "age")
)
}
\arguments{
\item{data_source}{Data.frame wich shuld be cleaned.}

\item{ommit_vars}{Vector with names of columns to ommit.}
}
\description{
Subset data to only contain community (taxa)
}
\keyword{internal}
