% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_into_proportions.r
\name{transform_into_proportions}
\alias{transform_into_proportions}
\title{Transform community data into proportions}
\usage{
transform_into_proportions(
  data_source_trans,
  sel_method = c("proportions", "percentages"),
  verbose = FALSE
)
}
\arguments{
\item{data_source_trans}{Data.frame with \code{label}, \code{res_age}, and all community data}

\item{sel_method}{variable to select result as either proportions (\code{percentages}) or
percentage (\code{percentages}).}

\item{verbose}{Logical. Should additional information be output?}
}
\description{
Tranform pollen data into proportions (or percentages)
}
\keyword{internal}
